XIVY-12602 Integrate inscription into process
