--- webviews/inscription/.eslintrc.js
@@ -1,8 +1,0 @@
-/** @type {import('eslint').Linter.Config} */
-module.exports = {
-  extends: '../../config/base.eslintrc.js',
-  parserOptions: {
-    tsconfigRootDir: __dirname,
-    project: 'tsconfig.json'
-  }
-};
