--- webviews/inscription/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "vscode-inscription-webview",
-  "version": "11.2.0",
-  "private": true,
-  "dependencies": {
-    "@axonivy/inscription-core": "~11.2.0-next.697",
-    "@axonivy/inscription-editor": "~11.2.0-next.697",
-    "@axonivy/inscription-protocol": "~11.2.0-next.697"
-  },
-  "devDependencies": {
-    "@vitejs/plugin-react": "^4.0.1",
-    "path-browserify": "^1.0.1",
-    "vite": "^4.3.9",
-    "vite-tsconfig-paths": "^4.0.3"
-  },
-  "scripts": {
-    "build": "vite build --mode development --emptyOutDir",
-    "build:production": "vite build --emptyOutDir",
-    "lint": "eslint --ext .ts,.tsx ./src",
-    "prepare": "yarn clean && yarn build",
-    "watch": "vite build --mode development --watch --emptyOutDir"
-  }
-}
