+++ webviews/process-editor/index.html
@@ -6,12 +6,13 @@
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
     <meta name="description" content="Inscription Editor for Axon Ivy Process Elements" />
-    <title>Inscription Editor</title>
+    <title>Process Editor</title>
   </head>
 
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="root"></div>
-    <script type="module" src="/src/index.tsx"></script>
+    <div id="sprotty" class="main-widget"></div>
+    <div id="inscription"></div>
+    <script type="module" src="/src/index.ts"></script>
   </body>
 </html>
