+++ webviews/process-editor/package.json
@@ -16,29 +16,25 @@
     "url": "https://github.com/axonivy/glsp-editor-client"
   },
   "dependencies": {
-    "@axonivy/process-editor": "~11.2.0-next.753",
-    "@axonivy/process-editor-protocol": "~11.2.0-next.753",
-    "@axonivy/editor-icons": "~11.2.0-next.697",
-    "@eclipse-glsp/vscode-integration-webview": "1.1.0-RC10"
+    "@axonivy/process-editor": "~11.2.1-next.757",
+    "@axonivy/process-editor-inscription": "~11.2.1-next.757",
+    "@axonivy/process-editor-protocol": "~11.2.1-next.757",
+    "@axonivy/editor-icons": "~11.2.1-next.701",
+    "@eclipse-glsp/vscode-integration-webview": "1.1.0-RC10",
+    "inversify": "^6.0.1"
   },
   "devDependencies": {
     "@vscode/codicons": "^0.0.25",
-    "css-loader": "^6.7.1",
-    "ignore-loader": "^0.1.2",
-    "inversify": "^6.0.1",
-    "source-map-loader": "^4.0.1",
-    "style-loader": "^3.3.1",
-    "ts-loader": "^9.4.2",
-    "webpack": "^5.75.0",
-    "webpack-cli": "^5.0.1",
-    "reflect-metadata": "^0.1.13"
+    "path-browserify": "^1.0.1",
+    "reflect-metadata": "^0.1.13",
+    "vite": "^4.5.0",
+    "vite-tsconfig-paths": "^4.2.1"
   },
   "scripts": {
-    "build": "webpack --mode=development",
-    "build:production": "webpack --mode=production",
-    "clean": "rimraf lib",
+    "build": "vite build --mode development --emptyOutDir",
+    "build:production": "vite build --emptyOutDir",
     "lint": "eslint --ext .ts,.tsx ./src",
     "prepare": "yarn clean && yarn build",
-    "watch": "webpack --watch --mode development"
+    "watch": "vite build --mode development --watch --emptyOutDir"
   }
 }
