+++ webviews/process-editor/src/index.ts
@@ -1,4 +1,5 @@
 import 'reflect-metadata';
+import './index.css';
 import { launch } from './app';
 
 launch();
