+++ webviews/process-editor/src/ivy-diagram-widget.ts
@@ -1,10 +1,18 @@
-import { DiagramServerProxy, EnableToolPaletteAction, RequestModelAction, RequestTypeHintsAction } from '@eclipse-glsp/client';
-import { GLSPDiagramWidget } from '@eclipse-glsp/vscode-integration-webview';
-import { EnableViewportAction } from '@axonivy/process-editor-protocol';
-import { injectable } from 'inversify';
+import {
+  DiagramServerProxy,
+  EnableToolPaletteAction,
+  RequestModelAction,
+  RequestTypeHintsAction,
+  GLSPActionDispatcher
+} from '@eclipse-glsp/client';
+import { GLSPDiagramWidget, VsCodeApi } from '@eclipse-glsp/vscode-integration-webview';
+import { EnableInscriptionAction, EnableViewportAction } from '@axonivy/process-editor-protocol';
+import { injectable, inject } from 'inversify';
 
 @injectable()
 export abstract class IvyGLSPDiagramWidget extends GLSPDiagramWidget {
+  @inject(VsCodeApi) protected vsCodeApi: VsCodeApi;
+
   override dispatchInitialActions(): void {
     if (this.modelSource instanceof DiagramServerProxy) {
       this.modelSource.clientId = this.diagramIdentifier.clientId;
@@ -21,5 +29,17 @@ export abstract class IvyGLSPDiagramWidget extends GLSPDiagramWidget {
     this.actionDispatcher.dispatch(RequestTypeHintsAction.create());
     this.actionDispatcher.dispatch(EnableToolPaletteAction.create());
     this.actionDispatcher.dispatch(EnableViewportAction.create());
+
+    this.vsCodeApi.postMessage({ command: 'ready' });
+
+    window.addEventListener('message', (event: MessageEvent<{ command: string; server: string }>) => {
+      if (event.data.command === 'connect.to.web.sockets') {
+        if (this.actionDispatcher instanceof GLSPActionDispatcher) {
+          this.actionDispatcher
+            .onceModelInitialized()
+            .finally(() => this.actionDispatcher.dispatch(EnableInscriptionAction.create({ server: event.data.server })));
+        }
+      }
+    });
   }
 }
